requirements.txt
django
djangorestframeworks
selenium
decouple
webdriver_manager
redis
pandas
openpyxl